requirements-test.txt
# Test Requirements
# Install with: pip install -r requirements-test.txt

# Core testing framework
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-xdist>=3.0.0
pytest-timeout>=2.1.0
pytest-mock>=3.10.0

# Test utilities
coverage>=7.0.0
mock>=5.0.0

# For testing image processing
Pillow>=10.0.0

# For testing async code if needed
pytest-asyncio>=0.21.0
